// src/app/model/user.model.ts
export class User {
  username: string;
  usercode: string;
  firstName: string;
  lastName: string;
  address: string;
  phone: string;
  email: string;
  avatarUrl: string;
  id: number;
  WarehouseReceive: number;
  userCode: string;
  lsUserSetting: any[];
  captchaImageLink: string;
  captchaCode: string;
  referralCode: string;

  // update
  createdDate?: Date;
  FirstName: string;
  LastName: string;

  country: string;
  countryDisplay: string;

  city: string;
  cityDisplay: string;

  district: string;
  districtDisplay: string;

  ward: string;
  wardDisplay: string;

  userId: number;

  Phone: string;
  Email: string;
  ChangeUser: number;
  Birthday: Date;
  birthday: Date;
  Sex: string;
  Chatname: string;
  Username: string;
  Password: string;
  Status: string;
  StatusDisplay: string;
  RightId: number;
  RightDisplay: string;
  LastChangePass: Date;
  Type: string;
  AvatarUrl: string;
  UserLevel: string;

  password: string;
  rePassword: string;
  lastChangePass: any;
  userLevel: string;
  rightId: number;
  warehouseReceive: number;
}